Move SliceZone component directly under /components in order to follow new Prismic-Nuxt Slice Machine project source organization
