+++ components/SliceZone.vue
